requirements.txt
Flask
python-dotenv
requests
pydub
simpleaudio
openai
gTTS
pydub
google-generativeai